% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arabia-package.R
\docType{package}
\name{arabia}
\alias{arabia}
\alias{arabia-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/arabia}
\item BugReports: \url{https://gitlab.com/hrbrmstr/arabia/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
