% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sl2.R
\name{read_sl2}
\alias{read_sl2}
\title{Read Lowrance binary track files}
\usage{
read_sl2(path, verbose = TRUE)
}
\arguments{
\item{path}{path to file}

\item{verbose}{if \code{TRUE} shows progress}
}
\value{
data frame (tibble)
}
\description{
Lowrance chart plotters use an ugly but straightforward binary format to encode
their tracks. This function reads "SL2" files and likely reads "SLG" & "SL3" files
as well.
}
\examples{
read_sl2(system.file("exdat", "example.sl2", package="arabia"))
}
\references{
\itemize{
\item \url{http://www.lowrance.com/}
\item \url{https://wiki.openstreetmap.org/wiki/SL2}
\item \url{https://github.com/kmpm/node-sl2format} (very helpful additional field breakdown)
\item \url{https://stackoverflow.com/q/52280751/1457051}
}
}
