% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arabia-package.R
\docType{package}
\name{arabia}
\alias{arabia}
\alias{arabia-package}
\title{Tools to Read 'Lowrance' Binary Track Files}
\description{
Lowrance' (\url{http://www.lowrance.com/}) chart plotters use an ugly
but straightforward binary format to encode their tracks. Tools are provided
to read 'SL2' files (and, very likely 'SLG' and 'SL3' files, too).
}
\details{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/arabia}
\item BugReports: \url{https://gitlab.com/hrbrmstr/arabia/issues}
}
}
\references{
\itemize{
\item \url{http://www.lowrance.com/}
\item \url{https://wiki.openstreetmap.org/wiki/SL2}
\item \url{https://github.com/kmpm/node-sl2format} (very helpful additional field breakdown)
\item \url{https://stackoverflow.com/q/52280751/1457051}
}
}
\author{
Bob Rudis (bob@rud.is)
}
